// src/core/shared/interfaces/formData.ts
export interface DiscussTheProjectInterface {
  name: {
    error: boolean | string;
    step: number;
    value: string;
  };
  company: {
    error: boolean | string;
    step: number;
    value: string;
  };
  email: {
    error: boolean | string;
    step: number;
    value: string;
  };
  skype: {
    error: boolean | string;
    step: number;
    value: string;
  };
  task: {
    error: boolean | string;
    step: number;
    value: string;
  };
  projectType: {
    error: boolean | string;
    step: number;
    value: string;
  };
  budget: {
    error: boolean | string;
    step: number;
    value: string;
  };
}